feat: tmdb api test(top_rated)

+++ src/main/java/com/movie/MovieReview/movie/dto/MovieDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class MovieDto {
-    private int id; //movie id
+    private Long id; //movie id
     private String title; //movie제목
     private String overview; //movie 줄거리
     private String poster_path; //movie 포스터 url https://image.tmdb.org/t/p/w500/poster_path
